Float to double replaced

+++ Renderer.cpp
@@ -11,11 +11,11 @@ Renderer::Renderer() {
 QImage Renderer::render(int imgWidth, int imgHeight) const {
     QImage img(imgWidth, imgHeight, QImage::Format_RGB32);
 
-    float aspectRatio = imgWidth / imgHeight;
+    double aspectRatio = imgWidth / imgHeight;
 
-    float focalLength = 1.0;
-    float viewportHeight = 2.0;
-    float viewportWidth =  aspectRatio * viewportHeight;
+    double focalLength = 1.0;
+    double viewportHeight = 2.0;
+    double viewportWidth =  aspectRatio * viewportHeight;
 
     Point3D origin(0, 0, 0);
     QVector3D horizontal(viewportWidth, 0, 0);
@@ -25,8 +25,8 @@ QImage Renderer::render(int imgWidth, int imgHeight) const {
     for (int ix = 0; ix < imgWidth; ix++) {
         for (int iy = imgHeight - 1; iy >= 0; iy--) {
             //Temporary simple render
-            float u = ix / float(imgWidth - 1);
-            float v = (imgHeight - iy - 1) / float(imgHeight - 1);
+            double u = ix / double(imgWidth - 1);
+            double v = (imgHeight - iy - 1) / double(imgHeight - 1);
 
             Ray ray(origin, bottomLeft + u * horizontal + v * vertical - origin);
             Color pixelColor = rayColor(ray);
@@ -38,7 +38,7 @@ QImage Renderer::render(int imgWidth, int imgHeight) const {
 
 Color Renderer::rayColor(const Ray& ray) const {
     QVector3D unitDirection = ray.direction().normalized();
-    float t = 0.5  * (unitDirection.y() + 1.0);
+    double t = 0.5  * (unitDirection.y() + 1.0);
     return (1.0 - t) * Color(1.0, 1.0, 1.0) + t * Color(0.5, 0.7, 1.0);
 }
